PR feedback.  Remove unsued code
