+++ src/commands/createStaticWebApp/IStaticWebAppWizardContext.ts
@@ -6,15 +6,12 @@
 import { WebSiteManagementClient, WebSiteManagementModels } from '@azure/arm-appservice';
 import { ICreateChildImplContext, IResourceGroupWizardContext } from 'vscode-azureextensionui';
 import { BranchData, ListOrgsForUserData, OrgForAuthenticatedUserData, RepoData } from '../../gitHubTypings';
-import { CreateScenario } from './CreateScenarioListStep';
 
 // creating a dummy repoData/branchData would be an annoying amount of work, so use this type to recognize when users have selected create new repo
 export type CreateNewResource = { name?: string; html_url?: string };
 export interface IStaticWebAppWizardContext extends IResourceGroupWizardContext, ICreateChildImplContext {
     accessToken: string;
     client: WebSiteManagementClient;
-
-    createScenario?: CreateScenario;
 
     orgData?: OrgForAuthenticatedUserData | ListOrgsForUserData;
     repoData?: RepoData | CreateNewResource;
